fix : commentservice 예외처리 수정

+++ src/main/java/sparat/spartaclone/comment/service/CommentService.java
@@ -77,6 +77,14 @@ public class CommentService {
 
         Optional<CommentLike> commentLike = commentLikeRepository.findByUserIdAndCommentId(user.getId(), commentId);
 
+        if(!commentLike.get().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("유저가 일치하지 않습니다.");
+        }
+
+        if(!user.getId().equals(comment.getUser().getId())) {
+            throw new EntityNotFoundException("유저가 일치하지 않습니다.");
+        }
+
         comment.updateComment(commentId, requestDto.getContent());
         return new CommentResponseDto(comment, commentLike.isPresent());
     }
@@ -90,6 +98,10 @@ public class CommentService {
         User user = userRepository.findByUsername(username).orElseThrow(
                 () -> new EntityNotFoundException(ErrorMessage.WRONG_USERNAME.getMessage())
         );
+
+        if(!user.getId().equals(comment.getUser().getId())) {
+            throw new EntityNotFoundException("유저가 일치하지 않습니다.");
+        }
 
         commentRepository.deleteById(commentId);
     }
